release: v1.17.2
Feat
1. 查看文件详情时能更准确知道该文件属于什么电视剧是否有匹配详情等信息

Bugfix
1. 删除指定云盘内指定季所有解析剧集时将该云盘内所有解析剧集都删除了的问题
2. 无法将已完成的资源重新创建一个同步任务的问题
3. 重命名云盘文件时将待索引文件名字也更新了的问题
4. 无法通过重命名重置掉已有电视剧解析结果的问题
5. 搜索电影时接口超时等错误会导致任务无法继续执行的问题

+++ pages/api/admin/source/delete.ts
@@ -1,5 +1,5 @@
 /**
- * @file 删除指定视频源（不删除文件）
+ * @file 管理后台/删除指定视频源（不删除文件）
  */
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from "next";
@@ -12,7 +12,11 @@ import { store } from "@/store";
 export default async function handler(req: NextApiRequest, res: NextApiResponse<BaseApiResp<unknown>>) {
   const e = response_error_factory(res);
   const { authorization } = req.headers;
-  const { drive_id } = req.query as Partial<{ drive_id: string }>;
+  const { drive_id, season_id, movie_id } = req.query as Partial<{
+    drive_id: string;
+    season_id: string;
+    movie_id: string;
+  }>;
   const t_res = await User.New(authorization, store);
   if (t_res.error) {
     return e(t_res);
@@ -20,18 +24,27 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse<
   if (!drive_id) {
     return e(Result.Err("缺少云盘 id"));
   }
+  if (!season_id && !movie_id) {
+    return e(Result.Err("不能删除云盘内全部解析结果"));
+  }
   const user = t_res.data;
-  await store.prisma.parsed_episode.deleteMany({
-    where: {
-      drive_id,
-      user_id: user.id,
-    },
-  });
-  await store.prisma.parsed_movie.deleteMany({
-    where: {
-      drive_id,
-      user_id: user.id,
-    },
-  });
+  if (season_id) {
+    await store.prisma.parsed_episode.deleteMany({
+      where: {
+        season_id,
+        drive_id,
+        user_id: user.id,
+      },
+    });
+  }
+  if (movie_id) {
+    await store.prisma.parsed_movie.deleteMany({
+      where: {
+        movie_id,
+        drive_id,
+        user_id: user.id,
+      },
+    });
+  }
   res.status(200).json({ code: 0, msg: "删除成功", data: null });
 }
